\name{manhplot-package}
\alias{manhplot-package}
\alias{manhplot}
\docType{package}
\title{
\packageTitle{manhplot}
}
\description{
\packageDescription{manhplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{manhplot}
\packageIndices{manhplot}
}
\author{
\packageAuthor{manhplot}

Maintainer: \packageMaintainer{manhplot}
}
\keyword{ package }
