% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.heatmap.v1.R
\name{manhplot}
\alias{manhplot}
\title{Generate the manhattan++ plot}
\usage{
manhplot(infile, outfile, configfile, snpfile, drawastiff = F,
  GWS = 5e-08, FDR = 0.001, MAF = 0.05, chrname = "chr",
  posname = "pos", pvalname = "pvalue", frqname = "maf",
  conseqname = "conseq")
}
\arguments{
\item{infile}{Input GWAS summary statistics}

\item{outfile}{Output file prefix for the manhattan++ plot}

\item{configfile}{Configuration file}

\item{snpfile}{Table of SNPs to visualize}

\item{drawastiff}{If TRUE draw a Tiff file, if FALSE draw a PDF file}

\item{GWS}{Genome wise significance pvalue threshold (5E-8 by default)}

\item{FDR}{False discovery Rate pvalue threshold (1E-3 by default)}

\item{MAF}{Minor Allele Frequency threshold}

\item{chrname}{Column name for chromosome in GWAS infile}

\item{posname}{Column name for position in GWAS infile}

\item{pvalname}{Column name for pvalue in GWAS infile}

\item{frqname}{column name for allele frequency in GWAS infile}

\item{conseqname}{column name for variant annotation consequence in GWAS infile}
}
\description{
Generate the manhattan++ plot
}
