\name{manhplot-package}
\alias{manhplot-package}
\alias{manhplot}
\docType{package}
\title{
\packageTitle{manhplot}
}
\description{
\packageDescription{manhplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{manhplot}
\packageIndices{manhplot}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{manhplot}

Maintainer: \packageMaintainer{manhplot}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
