% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.heatmap.v1.R
\name{manhplot}
\alias{manhplot}
\title{Generate the manhattan++ plot}
\usage{
manhplot(infile, outfile, configfile, snpfile, drawastiff = F,
  GWS = 5e-08, FDR = 0.001, MAF = 0.05, chrname = "chr",
  posname = "pos", pvalname = "pvalue", frqname = "maf",
  conseqname = "conseq", showgenes = F, showrsids = F,
  pos.split = 3e+06, pval.split = 0.125, max.pval = 20)
}
\arguments{
\item{infile}{Input GWAS summary statistics}

\item{outfile}{Output file prefix for the manhattan++ plot}

\item{configfile}{Configuration file}

\item{snpfile}{Table of SNPs to visualize}

\item{drawastiff}{If TRUE draw a Tiff file, if FALSE draw a PDF file}

\item{GWS}{Genome wise significance pvalue threshold (5E-8 by default)}

\item{FDR}{False discovery Rate pvalue threshold (1E-3 by default)}

\item{MAF}{Minor Allele Frequency threshold}

\item{chrname}{Column name for chromosome in GWAS infile}

\item{posname}{Column name for position in GWAS infile}

\item{pvalname}{Column name for pvalue in GWAS infile}

\item{frqname}{column name for allele frequency in GWAS infile}

\item{conseqname}{column name for variant annotation consequence in GWAS infile}

\item{showgenes}{If T shows known genes as bubbles on main manhattan plot, if F show positions of interest as bubbles}

\item{showrsids}{If showgenes is T, then show the rsids, rather than genes}
}
\description{
Generate the manhattan++ plot
}
\examples{
library(manhplot)
## Load R.utils for gzip functionality
install.packages("R.utils")
library(R.utils)

## unzip the data included with this package
gunzip(system.file("extdata","cad.add.160614_manhformatv3.txt.gz",package = "manhplot"))

infile<-system.file("extdata","cad.add.160614_manhformatv3.txt",package = "manhplot")
configfile<-system.file("extdata","config.txt", package = "manhplot")
snpfile<-system.file("extdata","56cad.add.160614.variants.txt", package = "manhplot")

manhplot(infile = infile,outfile = "default-plot", configfile = configfile, snpfile = snpfile)

## zip the data
gzip(system.file("extdata","cad.add.160614_manhformatv3.txt",package = "manhplot"))

}
